@@ -6,23 +6,16 @@ author: Ravi Teja Gudapati <tejainece@gmail.com>
 documentation:
 
 environment:
-  sdk: '>=1.20.1 <2.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
-  meta: ^1.1.2
-  grizzly_array: ^0.4.7
-  grizzly_linalg: ^0.4.2
+  grizzly:
+    path: ../grizzly
+  grizzly_linalg:
+    path: ../grizzly_linalg
+  meta: ^1.8.0
 
 dev_dependencies:
   test:
 
 dev_dependency_overrides:
-  grizzly_array:
-    path:
-      ../grizzly_array
-  grizzly_linalg:
-    path:
-      ../grizzly_linalg
-  grizzly_primitives:
-    path:
-      ../grizzly_primitives
