@@ -3,7 +3,7 @@ part of grizzly.regress;
 /// Mixin for regression model
 abstract class RegressionModel {
   /// Estimated coefficients for the linear regression problem
-  Double1DFix get coeff;
+  Iterable<num> get coeff;
 
   /// Is the intercept induced by the model?
   bool get interceptFitted;
@@ -11,7 +11,7 @@ abstract class RegressionModel {
   /// Independent term in the linear model
   double get intercept {
     if (interceptFitted) {
-      return coeff.last;
+      return coeff.last.toDouble();
     } else {
       return 0.0;
     }
@@ -19,30 +19,30 @@ abstract class RegressionModel {
 
   /// Predicts y_hat for given single independent variable sample [x] and
   /// coefficient [coeff]
-  double predictOne(Numeric1DView x) {
+  double predictOne(Iterable<num> x) {
     if (interceptFitted) {
       if ((x.length + 1) != coeff.length)
-        throw new ArgumentError.value(
+        throw ArgumentError.value(
             "Have ${coeff.length} coefficients (with intercept fitted). But number of exogeneous variables is ${x.length}. ${coeff.length - 1} exogenous variables expected!");
 
-      double ret = coeff.last;
+      double ret = coeff.last.toDouble();
 
-      ret += coeff.slice(0, coeff.length - 1).dot(x); // TODO use sliceView
+      ret += coeff.cut(0, coeff.length - 1).dot(x);
 
       return ret;
     } else {
       if (x.length != coeff.length)
-        throw new ArgumentError(
+        throw ArgumentError(
             "Have ${coeff.length} coefficients. But number of exogeneous variables is ${x.length}! ${coeff.length} exogenous variables expected!");
 
-      return coeff.dot(x);
+      return coeff.dot(x).toDouble();
     }
   }
 
   /// Predicts y_hat for given multiple independent variable samples [x] and
   /// coefficient [coeff]
-  Double1D predict(Numeric2DView x) {
-    final ret = new Double1D.sized(x.numRows);
+  Double1D predict(Num2DView x) {
+    final ret = Double1D.filled(x.numRows, 0);
     for (int i = 0; i < x.numRows; i++) {
       ret[i] = predictOne(x[i]);
     }
