@@ -5,7 +5,7 @@ library grizzly.regress;
 
 import 'dart:math' as math;
 import 'package:meta/meta.dart';
-import 'package:grizzly_array/grizzly_array.dart';
+import 'package:grizzly/grizzly.dart';
 import 'package:grizzly_linalg/grizzly_linalg.dart';
 import 'package:text_table/text_table.dart';
 
@@ -22,16 +22,16 @@ part 'logistic/logistic.dart';
 /// Defines interface for regression
 abstract class Regression {
   /// Fit simple model with one independent variable
-  RegressionResult fitOne(Numeric1D x, Numeric1D y, {bool fitIntercept: false});
+  RegressionResult fitOne(Num1D x, Num1D y, {bool fitIntercept: false});
 
   /// Fit model with multiple independent variable
-  RegressionResult fit(Numeric2D x, Numeric1D y, {bool fitIntercept: false});
+  RegressionResult fit(Num2D x, Num1D y, {bool fitIntercept: false});
 
   /// Fit simple model with one independent variable
-  List<RegressionResult> fitOneMultivariate(Numeric1D x, Numeric2D y,
+  List<RegressionResult> fitOneMultivariate(Iterable<num> x, Num2DView y,
       {bool fitIntercept: false});
 
   /// Fit model with multiple independent variable
-  List<RegressionResult> fitMultivariate(Numeric2D x, Numeric2D y,
+  List<RegressionResult> fitMultivariate(Num2D x, Num2D y,
       {bool fitIntercept: false});
 }
