@@ -1,5 +1,5 @@
 import 'package:grizzly_regress/grizzly_regress.dart';
-import 'package:grizzly_array/grizzly_array.dart';
+import 'package:grizzly/grizzly.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -7,7 +7,7 @@ void main() {
     setUp(() {});
 
     test('First Test', () {
-      final y = new String1D([
+      final y = [
         'zero',
         'zero',
         'one',
@@ -29,8 +29,8 @@ void main() {
         'three',
         'five',
         'three',
-      ]);
-      final classes = new MakeClasses.label(y);
+      ];
+      final classes = MakeClasses.label(y);
       expect(classes.labels, ['zero', 'one', 'two', 'three', 'five']);
       expect(classes.counts,
           {'zero': 3, 'one': 5, 'two': 3, 'three': 5, 'five': 5});
